sc - move files into the main folder of javascript/src/
