+++ javascript/src/index.js
@@ -1,8 +1,8 @@
 import React from "react";
 import ReactDOM from "react-dom";
 import "./index.css";
-import App from "./App";
-import Auth0ProviderWithHistory from "components/Auth/Auth0ProviderWithHistory";
+import App from "main/App";
+import Auth0ProviderWithHistory from "main/components/Auth/Auth0ProviderWithHistory";
 import { BrowserRouter } from "react-router-dom";
 import * as serviceWorker from "./serviceWorker";
 import "bootstrap/dist/css/bootstrap.css";
