+++ javascript/src/main/components/Auth/PrivateRoute.js
@@ -1,7 +1,7 @@
 import React from "react";
 import { Route } from "react-router-dom";
 import { withAuthenticationRequired } from "@auth0/auth0-react";
-import Loading from "components/Loading/Loading";
+import Loading from "main/components/Loading/Loading";
 
 const PrivateRoute = ({ component, ...args }) => (
   <Route
