+++ javascript/src/main/components/Footer/AppFooter.js
