+++ javascript/src/main/components/Footer/AppFooter.test.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { render } from "@testing-library/react";
-import AppFooter from "./AppFooter";
+import AppFooter from "main/components/Footer/AppFooter";
 
 describe("AppFooter tests", () => {
   test("renders without crashing", () => {
