+++ javascript/src/main/components/Nav/AuthNav.test.js
